+++ src/components/TabBarImage.tsx
@@ -1,0 +1,11 @@
+import React from "react"
+import { Image, ImageSourcePropType } from "react-native"
+
+
+export default function TabBarImage({ source }: { source: ImageSourcePropType }) {
+
+    return (
+        <Image testID="tabIconImage" style={{ marginHorizontal: -7, marginVertical: -4 }} source={source} />
+    )
+
+}